Move from IntPtr to nint to avoid runtime casting errors. (#5791)
We still had a number of potential and actual overflow scenarios. In order to make this more resilient, move SendMessage, PostMessage, and Get/SetWindowLong from IntPtr to nint.

This also makes things slightly faster as we aren't creating IntPtr structs and invoking the cast methods on it.

Note that IntPtr implicitly converts to nint. I've removed most of the actual IntPtr usages. I've also removed most of the HandleRef overloads and vetted the callers to do the right thing in regards to avoiding finalizer issues.

There is more we can do here, but this is a big first step.

Also clean up DateTimePicker and move SYSTEMTIME converters to SYSTEMTIME.

I've tried to remove all (IntPtr) casts in SendMessage calls.

+++ src/System.Windows.Forms.Primitives/tests/UnitTests/System/Windows/Forms/DeviceContextScopeTests.cs
@@ -84,7 +84,7 @@ namespace System.Windows.Forms.Tests
             try
             {
                 // We get the same HDC out
-                Assert.Equal(dcScope.HDC.Handle, hdc);
+                Assert.Equal(dcScope.HDC.Handle, (nint)hdc);
                 current = Gdi32.GetCurrentObject(dcScope, Gdi32.OBJ.BRUSH);
                 Assert.Equal(blueBrush.HBrush.Handle, current.Handle);
                 Gdi32.SelectObject(dcScope, redBrush);
